Implementation of better logic for Back button in MovieView
Now, if a user click on a movie and then on the Back button, the Back button will bring the user back to his previous page (not automatically in Home). For exemple, a user accessing movie from his profile page will be brought back to his profile page instead of Home page.